fix(trace): move to folder
